requirements.txt
alembic==1.0.3
aniso8601==4.0.1
beautifulsoup4==4.7.1
cachetools==3.1.0
certifi==2018.10.15
chardet==3.0.4
Click==7.0
colorama==0.4.0
config==0.4.0
Flask==1.0.2
Flask-Cors==3.0.7
Flask-JWT-Extended==3.24.0
flask-marshmallow==0.9.0
Flask-Migrate==2.3.0
Flask-RESTful==0.3.6
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
google==2.0.1
google-api-python-client==1.7.8
google-auth==1.6.2
google-auth-httplib2==0.0.3
gunicorn==19.9.0
httpie==1.0.3
httplib2==0.12.0
idna==2.7
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.0.7
MarkupSafe==1.1.0
marshmallow==2.18.0
marshmallow-sqlalchemy==0.16.0
mysql-connector==2.2.9
passlib==1.7.1
pyasn1==0.4.5
pyasn1-modules==0.2.4
Pygments==2.4.2
PyJWT==1.7.1
python-dateutil==2.7.5
python-editor==1.0.3
pytz==2018.7
requests==2.22.0
rsa==4.0
six==1.11.0
soupsieve==1.7.3
SQLAlchemy==1.3.5
sqlautocode==0.7
uritemplate==3.0.0
urllib3==1.25.3
Werkzeug==0.15.5
